Use layout defined as reference in the API response. Dynamically update as well
