+++ src/core/mixins/ComponentMixin.js
@@ -1,8 +1,20 @@
+import consola from "consola"
 import ResourceMixin from "@cwa/nuxt-module/core/mixins/ResourceMixin.js"
-import {StoreCategories} from "@cwa/nuxt-modulecore/storage";
+import { StoreCategories } from "@cwa/nuxt-modulecore/storage"
 
 const category = StoreCategories.Component
 
 export default {
-  mixins: [ResourceMixin(category)]
+  mixins: [ResourceMixin],
+  computed: {
+    resource() {
+      const type = this.$cwa.$storage.getTypeFromIri(this.iri, category)
+      if (!type) {
+        consola.warn(`Could not resolve a resource type for iri ${this.iri} in the category ${category}`)
+        return null
+      }
+      consola.info(`Resolved resource type for iri ${this.iri} in the category ${category} to ${type}`)
+      return this.$cwa.resources[type].byId[this.iri]
+    }
+  }
 }
