+++ src/core/mixins/PageMixin.js
@@ -1,8 +1,19 @@
 import ResourceMixin from "@cwa/nuxt-module/core/mixins/ResourceMixin.js"
-import {StoreCategories} from "@cwa/nuxt-modulecore/storage";
-
-const category = StoreCategories.PageData
 
 export default {
-  mixins: [ResourceMixin(category)]
+  mixins: [ResourceMixin],
+  computed: {
+    resource() {
+      return this.$cwa.resources.Page.byId[this.iri]
+    },
+    layout() {
+      return this.$cwa.resources.Layout.byId[this.resource.layout]
+    }
+  },
+  watch: {
+    'layout.reference': {
+      handler() { return this.$cwa.setLayout(this.layout.reference) },
+      immediate: true
+    }
+  }
 }
