+++ src/core/mixins/ResourceMixin.js
@@ -1,7 +1,6 @@
-import consola from 'consola'
 import ComponentCollection from '@cwa/nuxt-module/core/templates/component-collection.vue'
 
-export default (category) => ({
+export default {
   props: {
     iri: {
       type: String,
@@ -10,15 +9,5 @@ export default (category) => ({
   },
   components: {
     ComponentCollection
-  },
-  computed: {
-    resource() {
-      const type = this.$cwa.$storage.getTypeFromIri(this.iri, category)
-      if (!type) {
-        consola.warn(`Could not resolve a resource type for iri ${this.iri}`)
-        return null
-      }
-      return this.$cwa.resources[type].byId[this.iri]
-    }
   }
-})
+}
